refactor(compiler): refactor function analyzer
