+++ kaba_compiler/src/semantic/function.rs
@@ -1,112 +1,5 @@
-use super::{
-    body,
-    error::{Result, SemanticError, SemanticErrorVariant},
-    state::AnalyzerState,
-    tn,
-    types::Type,
-};
-use crate::{
-    ast::{AstNode, FunctionParam},
-    AstNodeVariant,
-};
-
-pub fn analyze_declaration(state: &AnalyzerState, node: &AstNode) -> Result<()> {
-    analyze_params_tn(state, node)?;
-    analyze_return_tn(state, node)?;
-
-    save_function_t(state, node)
-}
-
-pub fn analyze_definition(state: &AnalyzerState, node: &AstNode) -> Result<Type> {
-    let (params_t, return_t) = get_function_t(state, node).unwrap_callable();
-
-    state.with_function_scope(node.id, return_t.clone(), || {
-        let params = node.variant.params().iter().zip(params_t);
-
-        for (param, t) in params {
-            let sym = &param.sym;
-            let sym_str = sym.variant.unwrap_symbol();
-
-            if !state.can_save_sym(sym_str) {
-                return Err(SemanticError {
-                    variant: SemanticErrorVariant::SymbolAlreadyExist(sym_str.to_string()),
-                    span: sym.span.clone(),
-                });
-            }
-
-            state.save_entity(sym.id, sym_str, t);
-        }
-
-        // Analyze function body
-        //
-        // We do this last in order to accommodate features such as
-        // recursive function call.
-
-        let body_t = body::analyze(state, node)?;
-
-        if !return_t.is_void() && body_t.is_void() {
-            return Err(SemanticError {
-                variant: SemanticErrorVariant::ReturnTypeMismatch {
-                    expected: return_t,
-                    get: Type::Void,
-                },
-                span: node.variant.sym().span.clone(),
-            });
-        }
-
-        Ok(())
-    })?;
-
-    Ok(Type::Void)
-}
-
-fn analyze_params_tn(state: &AnalyzerState, node: &AstNode) -> Result<()> {
-    for FunctionParam { tn: param_tn, .. } in node.variant.params() {
-        tn::analyze(state, param_tn, false)?;
-    }
-    Ok(())
-}
-
-fn analyze_return_tn(state: &AnalyzerState, node: &AstNode) -> Result<()> {
-    if let Some(return_tn) = node.variant.return_tn() {
-        tn::analyze(state, return_tn, true)?;
-    }
-    Ok(())
-}
-
-fn save_function_t(state: &AnalyzerState, node: &AstNode) -> Result<()> {
-    let sym = node.variant.sym();
-    let sym_str = sym.variant.unwrap_symbol();
-
-    let params_t = node
-        .variant
-        .params()
-        .iter()
-        .map(|p| Type::from(&p.tn))
-        .collect();
-    let return_t = Box::new(node.variant.return_tn().map_or(Type::Void, Type::from));
-    let function_t = Type::Callable { params_t, return_t };
-
-    if !state.can_save_sym(sym_str) {
-        return Err(SemanticError {
-            variant: SemanticErrorVariant::SymbolAlreadyExist(String::from(sym_str)),
-            span: sym.span.clone(),
-        });
-    }
-
-    state.save_entity(sym.id, sym_str, function_t);
-
-    Ok(())
-}
-
-fn get_function_t(state: &AnalyzerState, node: &AstNode) -> Type {
-    if let AstNodeVariant::FunctionDefinition { sym, .. } = &node.variant {
-        let sym_string = &sym.variant.unwrap_symbol();
-        state.get_sym_t(sym_string).unwrap().unwrap_entity()
-    } else {
-        unreachable!()
-    }
-}
+pub mod declaration;
+pub mod definition;
 
 #[cfg(test)]
 mod tests {
